% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtests.R
\name{sdfra}
\alias{sdfra}
\title{Stochastic dominance tests for risk averters}
\usage{
sdfra(data1, data2, B = 2000)
}
\arguments{
\item{data1}{the distribution you want to compare with data2}

\item{data2}{the distribution you want to compare with data1}

\item{B}{the computing times for bootstrap}
}
\value{
a list that contains tests result for FSD, SSD, TSD and some statistics \code{output}
}
\description{
Stochastic dominance tests for risk averters.
}
\examples{
\dontrun{
    set.seed(666)
    data11 <- rnorm(200, 1, 1)
    data12 <- rnorm(100, 0, 1)
    test1 <- sdfra(data11, data12)
    print(test1$sd)

    data21 <- rnorm(100, 0, 2)
    data22 <- rnorm(150, 0, 1)
    test2 <- sdfra(data21, data22)
    print(test2$sd)
}
}
