% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtests.R
\name{sdfrs}
\alias{sdfrs}
\title{Stochastic dominance tests for risk seekers}
\usage{
sdfrs(data1, data2, B = 2000)
}
\arguments{
\item{data1}{the distribution you want to compare with data2}

\item{data2}{the distribution you want to compare with data1}

\item{B}{the computing times for bootstrap}
}
\value{
a list that contains tests result for FSD, SSD, TSD and some statistics \code{output}
}
\description{
Stochastic dominance tests for risk seekers.
}
\examples{
\dontrun{
    set.seed(666)
    data31 <- rnorm(200, 2, 1)
    data32 <- rnorm(100, 0, 1)
    test3 <- sdfrs(data31, data32)
    print(test3$sd)

    data41 <- rnorm(100, 0, 2)
    data42 <- rnorm(150, 0, 1)
    test4 <- sdfrs(data41, data42)
    print(test4$sd)
}
}
