\name{StatComp21039-package}
\alias{StatComp21039-package}
\docType{package}
\title{
  Provide several functions for the final homework of the 'Statistical Computing' course in 2021 fall
}
\description{
  Two R functions can be used to do the Stochastic dominance tests for risk averters and risk seekers.
}
\author{
LingFeng Qi
Maintainer: LingFeng Qi <qlf1999@mail.ustc.edu.cn>
}
\keyword{ package }
